Conflicts:
	app/src/main/java/net/ducksmanager/whattheduck/IssueList.java

+++ app/src/main/java/net/ducksmanager/whattheduck/RetrieveTask.java
@@ -30,7 +30,7 @@ public class RetrieveTask extends AsyncTask<Object, Object, String> {
     protected void onPostExecute(String response) {
         if (this.thrownException != null) {
             if (progressBarId != null) {
-                WhatTheDuck.wtd.toggleProgressbarLoading(progressBarId.intValue(), false);
+                WhatTheDuck.wtd.toggleProgressbarLoading(progressBarId, false);
             }
 
             if (this.thrownException instanceof SecurityException) {
